FW-1455, FW-1321, FW-1273 - simplify docker setup (#336)
* Simplify docker process

* Speed up and reogranize docker builds

* Run through README and modify accordingly

* Mods after testing full build

* Changes following Jenkins testing

* Cypress changes

* Update base url

* Update base url

* Try Cypress version on 4.8.0

* Update cypress + force CI file

* Update cypress + force CI file

* No need for default config file. probably relative to e2e

* Slightly better output in docker for cypress

* Upgrade testing library cypress

* Update deprecated cypress tools methods

* Add labels and teardown

* Update gitignore + docker private sample

* Teardown and local conf + README updates

* Address review issues

+++ docker/setup.sh
@@ -1,5 +1,14 @@
 #!/bin/sh
-nuxeoctl mp-install --accept=yes /opt/nuxeo/server/nxserver/tmp/FirstVoices-marketplace-package-latest.zip
+
+PREBUILT_ZIP=/opt/nuxeo/server/tmp/FirstVoices-marketplace-package-latest.zip
+
+
+if test -f "$PREBUILT_ZIP"; then
+    nuxeoctl mp-install --accept=yes /opt/nuxeo/server/tmp/FirstVoices-marketplace-package-latest.zip
+else
+    nuxeoctl mp-install --accept=yes /opt/nuxeo/server/nxserver/tmp/FirstVoices-marketplace-package-latest.zip
+fi
+
 touch /opt/nuxeo/server/nxserver/tmp/this.folder.is.mounted.in.docker.container
 
 # This will exec the CMD from your Dockerfile, i.e. "npm start"
